requirements.txt
Django==4.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
